Fix code quality issues
